+++ packages/adhere-ui-antdformitem/es/FormItemGeneratorToDict/CascaderMulitFormItem/index.d.ts
@@ -1,8 +1,0 @@
-/**
- * CascaderMulitFormItem
- * @param props
- * @return {JSX.Element}
- * @constructor
- */
-declare function CascaderMulitFormItem(props: any): JSX.Element;
-export default CascaderMulitFormItem;