+++ packages/adhere-ui-antdformitem/es/FormItemGeneratorToDict/Fields/Radio.d.ts
@@ -1,6 +1,0 @@
-declare const _default: () => {};
-/**
- * initRadio
- * @description 初始化Radio
- */
-export default _default;