+++ packages/adhere-ui-antdformitem/es/FormItemGeneratorToDict/Fields/Table.d.ts
@@ -1,6 +1,0 @@
-declare const _default: () => {};
-/**
- * initTable
- * @description 初始化Table
- */
-export default _default;