+++ packages/adhere-ui-antdformitem/lib/FormItemGeneratorToDict/Fields/CheckBox.d.ts
@@ -1,6 +1,0 @@
-declare const _default: () => {};
-/**
- * initCheckBox
- * @description 初始化CheckBox
- */
-export default _default;