+++ packages/adhere-ui-antdformitem/lib/FormItemGeneratorToDict/Fields/Tag.js
@@ -1,2 +1,2 @@
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var FormItemComponents={};exports.default=function(){return FormItemComponents};
+"use strict";require("core-js/modules/es.object.define-property.js"),Object.defineProperty(exports,"__esModule",{value:!0});var FormItemComponents={};exports.default=function(){return FormItemComponents};
 //# sourceMappingURL=Tag.js.map