+++ packages/adhere-ui-antdformitem/lib/FormItemGeneratorToDict/Fields/TreeSelect.d.ts
@@ -1,6 +1,0 @@
-declare const _default: () => {};
-/**
- * initTreeSelect
- * @description 初始化TreeSelect
- */
-export default _default;