+++ packages/adhere-ui-searchtable/lib/index.d.ts
@@ -53,7 +53,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -110,7 +110,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -167,7 +167,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -224,7 +224,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -281,7 +281,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -338,7 +338,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -395,7 +395,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -452,7 +452,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -509,7 +509,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -566,7 +566,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -623,7 +623,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -680,7 +680,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -737,7 +737,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -794,7 +794,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -851,7 +851,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;
@@ -908,7 +908,7 @@ declare const _default: {
             onSubTableChange(pagination: any, filters: any, sorter: any): void;
             onSearchPanelCollapseBefore(): void;
             unMountSearchAndPaginParamsDeal(): void;
-            initSearchAndPaginParams(): any;
+            initSearchAndPaginParams(queryReduce?: ((key: string, v: any) => any) | undefined): any;
             hasAdvancedSearch(): boolean;
             hasNumberColumnFixed(): boolean;
             hasOptionColumnFixed(): boolean;